third_party/blink/web_tests/platform/generic/fast/url/host-lowercase-per-scheme-expected.txt

Canonical URL should have a lowercase hostname for server based protocols.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS successfullyParsed is true

TEST COMPLETE
